next step

+++ libnd4j/include/pointercast.h
@@ -25,7 +25,7 @@
 
 typedef void* Nd4jPointer;
 typedef long long Nd4jLong;
-typedef unsigned long long Nd4jULong;
+typedef uint64_t Nd4jULong;
 typedef int Nd4jStatus;
 
 #define ND4J_STATUS_OK            0
